# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLMNightRun",
    "description": "멀티 LLM 통합 자동화 플랫폼",
    "version": "0.1.0",
    "x-logo": {
      "url": "/static/logo.png",
      "altText": "LLMNightRun"
    }
  },
  "paths": {
...
    "/agent/create": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Agent",
        "description": "새 에이전트를 생성합니다.\n\nReturns:\n    AgentResponse: 생성된 에이전트 정보",
        "operationId": "create_agent_agent_create_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "에이전트를 찾을 수 없음"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/agent/{agent_id}/run": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Run Agent",
        "description": "에이전트를 실행합니다.\n\nArgs:\n    agent_id: 에이전트 ID\n    request: 에이전트 실행 요청 정보\n    background_tasks: 백그라운드 작업 객체\n    db: 데이터베이스 세션\n    \nReturns:\n    AgentResponse: 에이전트 실행 상태\n    \nRaises:\n    HTTPException: 에이전트를 찾을 수 없는 경우",
        "operationId": "run_agent_agent__agent_id__run_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "에이전트를 찾을 수 없음"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/agent/{agent_id}/status": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent Status",
        "description": "에이전트의 현재 상태를 조회합니다.\n\nArgs:\n    agent_id: 에이전트 ID\n    \nReturns:\n    AgentResponse: 에이전트 상태 정보",
        "operationId": "get_agent_status_agent__agent_id__status_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "에이전트를 찾을 수 없음"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/agent/{agent_id}": {
      "delete": {
        "tags": [
          "agent"
        ],
        "summary": "Delete Agent",
        "description": "에이전트를 삭제합니다.\n\nArgs:\n    agent_id: 에이전트 ID\n    \nReturns:\n    dict: 삭제 결과 메시지",
        "operationId": "delete_agent_agent__agent_id__delete",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "에이전트를 찾을 수 없음"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "workspace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace"
          },
          "max_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AgentRequest",
        "description": "에이전트 요청 모델"
      },
      "AgentResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          },
          "messages": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Messages"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "state"
        ],
        "title": "AgentResponse",
        "description": "에이전트 응답 모델"
      },
      "AgentState": {
        "type": "string",
        "enum": [
          "idle",
          "running",
          "finished",
          "error"
        ],
        "title": "AgentState",
        "description": "에이전트 상태"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
    {
      "url": "",
      "description": "개발 서버"
    }
  ]
}